PR feedback (move resource_pool_workflow, misc factoring)
