+++ resource_pool/pool_client/__init__.py
@@ -1,1 +1,2 @@
 from .resource_pool_client import ResourcePoolClient
+from .resource_pool_workflow import ResourcePoolWorkflow
