+++ resource_pool/starter.py
@@ -4,7 +4,7 @@ from typing import Any
 from temporalio.client import Client, WorkflowFailureError, WorkflowHandle
 from temporalio.common import WorkflowIDConflictPolicy
 
-from resource_pool.resource_pool_workflow import (
+from resource_pool.pool_client.resource_pool_workflow import (
     ResourcePoolWorkflow,
     ResourcePoolWorkflowInput,
 )
