+++ resource_pool/worker.py
@@ -4,7 +4,7 @@ import logging
 from temporalio.client import Client
 from temporalio.worker import Worker
 
-from resource_pool.resource_pool_workflow import ResourcePoolWorkflow
+from resource_pool.pool_client.resource_pool_workflow import ResourcePoolWorkflow
 from resource_pool.resource_user_workflow import ResourceUserWorkflow, use_resource
